+++ src/assets/common/loading.gif
